Updating the model. Broke it out into a train model, html model, and base.
